# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: SsuBench API
  description: |
    REST API платформы SsuBench — сервис где заказчики размещают задания, 
    исполнители откликаются на них, а оплата выполненной работы производится виртуальными баллами.
    
    ## Основные сценарии использования
    
    ### Для заказчика
    1. Регистрация и вход
    2. Создание задачи
    3. Выбор исполнителя из откликов
    4. Подтверждение выполнения и оплата
    
    ### Для исполнителя
    1. Регистрация и вход
    2. Поиск доступных задач
    3. Отклик на задачу
    4. Выполнение и ожидание оплаты
    
    ### Для администратора
    1. Просмотр всех пользователей
    2. Блокировка/разблокировка нарушителей
    
    ## Аутентификация
    Все запросы кроме `/api/auth/**` требуют JWT токен в заголовке:
    ```
    Authorization: Bearer <ваш_токен>
    ```
  version: 1.0.0
  contact:
    name: SsuBench Team

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Регистрация, вход, обновление токенов
  - name: Tasks
    description: Управление задачами
  - name: Bids
    description: Отклики на задачи
  - name: Payments
    description: Подтверждение выполнения и оплата
  - name: Admin
    description: Администрирование пользователей

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации или пользователь уже существует

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Вход по логину и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный логин или пароль

  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление пары токенов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Невалидный refresh токен

  /api/tasks:
    post:
      tags: [Tasks]
      summary: Создать новую задачу
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Задача создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Ошибка валидации
        '403':
          description: Только заказчик может создавать задачи
    
    get:
      tags: [Tasks]
      summary: Получить все задачи с пагинацией
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseTask'

  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Получить задачу по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
    
    put:
      tags: [Tasks]
      summary: Обновить задачу
      description: Можно обновить только задачу в статусе OPEN
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Задача обновлена
        '400':
          description: Нельзя изменить задачу в работе
        '403':
          description: Только заказчик может изменять

  /api/tasks/{id}/cancel:
    post:
      tags: [Tasks]
      summary: Отменить задачу
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача отменена
        '400':
          description: Нельзя отменить выполненную задачу
        '403':
          description: Только заказчик может отменять

  /api/tasks/my:
    get:
      tags: [Tasks]
      summary: Мои задачи (для заказчика)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список задач заказчика

  /api/tasks/available:
    get:
      tags: [Tasks]
      summary: Доступные задачи для исполнителя
      description: Задачи в статусе OPEN, кроме задач текущего пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список доступных задач

  /api/tasks/{id}/complete:
    post:
      tags: [Payments]
      summary: Завершить задачу (исполнитель)
      description: Только выбранный исполнитель может завершить задачу
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача завершена
        '400':
          description: Задача не в статусе IN_PROGRESS
        '403':
          description: Только выбранный исполнитель может завершить

  /api/tasks/{id}/confirm:
    post:
      tags: [Payments]
      summary: Подтвердить выполнение и оплатить (заказчик)
      description: |
        Атомарная операция:
        1. Списывает баллы у заказчика
        2. Начисляет баллы исполнителю
        3. Создаёт запись о платеже
        4. Меняет статус задачи на CONFIRMED
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Выполнение подтверждено, оплата проведена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Недостаточно средств или задача не завершена
        '403':
          description: Только заказчик может подтвердить
        '409':
          description: Платеж уже существует

  /api/tasks/{id}/payment:
    get:
      tags: [Payments]
      summary: Получить информацию о платеже
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о платеже
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Платеж не найден

  /api/bids/task/{taskId}:
    post:
      tags: [Bids]
      summary: Создать отклик на задачу
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBidRequest'
      responses:
        '200':
          description: Отклик создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'
        '400':
          description: Нельзя откликнуться на свою задачу
        '403':
          description: Только исполнитель может откликаться
        '409':
          description: Вы уже откликнулись на эту задачу
    
    get:
      tags: [Bids]
      summary: Получить все отклики на задачу
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список откликов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseBid'

  /api/bids/{id}:
    get:
      tags: [Bids]
      summary: Получить отклик по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отклик найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'

  /api/bids/{id}/accept:
    post:
      tags: [Bids]
      summary: Принять отклик (заказчик)
      description: |
        После принятия:
        - Статус отклика меняется на ACCEPTED
        - Статус задачи меняется на IN_PROGRESS
        - Все остальные отклики отклоняются
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отклик принят
        '403':
          description: Только заказчик может принять
        '409':
          description: У задачи уже есть принятый исполнитель

  /api/bids/{id}/reject:
    post:
      tags: [Bids]
      summary: Отклонить отклик (заказчик)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отклик отклонён
        '400':
          description: Можно отклонить только PENDING отклик
        '403':
          description: Только заказчик может отклонить

  /api/bids/{id}/cancel:
    post:
      tags: [Bids]
      summary: Отменить отклик (исполнитель)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отклик отменён
        '400':
          description: Можно отменить только PENDING отклик
        '403':
          description: Только автор отклика может отменить

  /api/bids/my:
    get:
      tags: [Bids]
      summary: Мои отклики (для исполнителя)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список откликов исполнителя

  /api/admin/users:
    get:
      tags: [Admin]
      summary: Получить всех пользователей (администратор)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseUser'
        '403':
          description: Только администратор

  /api/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Получить пользователя по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /api/admin/users/{id}/block:
    post:
      tags: [Admin]
      summary: Заблокировать пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь заблокирован
        '400':
          description: Нельзя заблокировать себя
        '404':
          description: Пользователь не найден

  /api/admin/users/{id}/unblock:
    post:
      tags: [Admin]
      summary: Разблокировать пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь разблокирован

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 4
          example: john_doe
        password:
          type: string
          minLength: 8
          example: SecurePass123!
        role:
          type: string
          enum: [CUSTOMER, EXECUTOR, ADMIN]
          example: CUSTOMER

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: john_doe
        password:
          type: string
          example: SecurePass123!

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    CreateTaskRequest:
      type: object
      required: [description, reward]
      properties:
        description:
          type: string
          maxLength: 2000
          example: Разработать REST API для сервиса
        reward:
          type: number
          format: decimal
          minimum: 0.01
          example: 100.00

    UpdateTaskRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
        reward:
          type: number
          format: decimal
          minimum: 0.01

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        customerId:
          type: integer
        customerUsername:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, COMPLETED, CONFIRMED, CANCELLED]
        reward:
          type: number
          format: decimal
        acceptedBidId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateBidRequest:
      type: object
      required: [description]
      properties:
        description:
          type: string
          maxLength: 2000
          example: Готов выполнить за 2 дня, есть опыт в этой области

    BidResponse:
      type: object
      properties:
        id:
          type: integer
        executorId:
          type: integer
        executorUsername:
          type: string
        taskId:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
        createdAt:
          type: string
          format: date-time

    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        bidId:
          type: integer
        amount:
          type: number
          format: decimal
        processedAt:
          type: string
          format: date-time

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [CUSTOMER, EXECUTOR, ADMIN]
        balance:
          type: number
          format: decimal
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time

    PageResponseTask:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    PageResponseBid:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BidResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    PageResponseUser:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean